Break into two modules
